+++ tasks/terraform/src/main/java/com/walmartlabs/concord/plugins/terraform/Constants.java
@@ -28,19 +28,19 @@ public final class Constants {
     public static final String DEFAULT_ENV_KEY = "defaultEnv";
     public static final String DESTROY_KEY = "destroy";
     public static final String DIR_KEY = "dir";
-    public static final String DIR_OR_PLAN_KEY = "dirOrPlan";
     public static final String EXTRA_ENV_KEY = "extraEnv";
     public static final String EXTRA_VARS_KEY = "extraVars";
     public static final String GIT_SSH_KEY = "gitSsh";
     public static final String IGNORE_ERRORS_KEY = "ignoreErrors";
     public static final String MODULE_KEY = "module";
+    public static final String PLAN_KEY = "plan";
     public static final String RESULT_KEY = "result";
     public static final String SAVE_OUTPUT_KEY = "saveOutput";
     public static final String STATE_ID_KEY = "stateId";
     public static final String VERBOSE_KEY = "verbose";
 
     public static final String[] ALL_IN_PARAMS = {ACTION_KEY, BACKEND_KEY, DEBUG_KEY, DEFAULT_ENV_KEY, DESTROY_KEY,
-            DIR_KEY, DIR_OR_PLAN_KEY, EXTRA_ENV_KEY, EXTRA_VARS_KEY, GIT_SSH_KEY, IGNORE_ERRORS_KEY, MODULE_KEY,
+            DIR_KEY, EXTRA_ENV_KEY, EXTRA_VARS_KEY, GIT_SSH_KEY, IGNORE_ERRORS_KEY, MODULE_KEY, PLAN_KEY,
             RESULT_KEY, SAVE_OUTPUT_KEY, STATE_ID_KEY, VERBOSE_KEY};
 
     private Constants() {