+++ tasks/terraform/src/main/java/com/walmartlabs/concord/plugins/terraform/actions/PlanAction.java
@@ -43,7 +43,8 @@ public class PlanAction extends Action {
     private final boolean destroy;
     private final boolean verbose;
     private final Path workDir;
-    private final Path dirOrPlan;
+    private final Path dir;
+    private final Path plan;
     private final Map<String, Object> extraVars;
     private final Map<String, String> env;
     private final boolean ignoreErrors;
@@ -58,12 +59,17 @@ public class PlanAction extends Action {
         this.destroy = MapUtils.get(cfg, Constants.DESTROY_KEY, false, Boolean.class);
         this.verbose = MapUtils.get(cfg, Constants.VERBOSE_KEY, false, Boolean.class);
 
+        // the process' working directory (aka the payload directory)
         this.workDir = getPath(cfg, com.walmartlabs.concord.sdk.Constants.Context.WORK_DIR_KEY, null);
         if (!workDir.isAbsolute()) {
             throw new IllegalArgumentException("'workDir' must be an absolute path, got: " + workDir);
         }
 
-        this.dirOrPlan = getPath(cfg, Constants.DIR_OR_PLAN_KEY, workDir);
+        // the TF files directory
+        this.dir = getPath(cfg, Constants.DIR_KEY, workDir);
+
+        // a file, created by the plan action
+        this.plan = getPath(cfg, Constants.PLAN_KEY, null);
 
         this.extraVars = MapUtils.get(cfg, Constants.EXTRA_VARS_KEY, null, Map.class);
         this.ignoreErrors = MapUtils.get(cfg, Constants.IGNORE_ERRORS_KEY, false, Boolean.class);
@@ -72,12 +78,14 @@ public class PlanAction extends Action {
 
     public PlanResult exec(Terraform terraform, Backend backend) throws Exception {
         try {
-            init(ctx, workDir, dirOrPlan, !verbose, env, terraform, backend);
+            init(ctx, workDir, dir, !verbose, env, terraform, backend);
+
+            Path dirOrPlanAbsolute = workDir.resolve(plan != null ? plan : dir);
 
             Path varsFile = createVarsFile(workDir, objectMapper, extraVars);
             Path outFile = getOutFile(workDir);
 
-            Terraform.Result r = new PlanCommand(debug, destroy, workDir, workDir.resolve(dirOrPlan), varsFile, outFile, env).exec(terraform);
+            Terraform.Result r = new PlanCommand(debug, destroy, workDir, dirOrPlanAbsolute, varsFile, outFile, env).exec(terraform);
             switch (r.getCode()) {
                 case 0: {
                     return PlanResult.noChanges(r.getStdout(), workDir.relativize(outFile).toString());